+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## ChangeLog
 
+## 0.3.0
+
+* Added `ConfigSpansURLOverride` to facilitate setting the Trace Observer URL
+  for Infinite Tracing on the New Relic Edge.
+
 ## 0.2.0
 
 * The SDK will now check metrics for infinity and NaN.  Metrics with invalid